atuh ih

+++ database/migrations/2026_07_31_111015_add_komoditas_id_to_permintaans_table.php
@@ -9,15 +9,20 @@ return new class extends Migration
     /**
      * Run the migrations.
      */
-    public function up(): void
-    {
-        Schema::table('permintaans', function (Blueprint $table) {
+   public function up(): void
+{
+    Schema::table('permintaans', function (Blueprint $table) {
+        if (!Schema::hasColumn('permintaans', 'komoditas_id')) {
             $table->foreignId('komoditas_id')->nullable()->constrained('komoditas')->onDelete('cascade');
-            
-            // Hapus ->change() agar Laravel membuat ini sebagai kolom baru
+        }
+        if (!Schema::hasColumn('permintaans', 'namatanaman')) {
             $table->string('namatanaman', 30)->nullable();
+        }
+        if (!Schema::hasColumn('permintaans', 'komoditas')) {
             $table->string('komoditas', 20)->nullable();
-        });
+        }
+    });
+
     }
 
     /**
